Refactor into scenes
