+++ src/scenes/Users/StatsPopup/style.scss
